refactor mask tests file structure
