--- tests/aswwu/behaviors/mask/mask_subtests.py
@@ -1,5 +1,0 @@
-import tests.aswwu.behaviors.mask.mask_requests as mask_requests
-import tests.utils as utils
-from tests.aswwu.data.paths import USERS_PATH
-import json
-from requests import Session, cookies
