+++ tests/aswwu/behaviors/mask/test_mask.py
@@ -1,123 +1,11 @@
-from tests.conftest import testing_server
+from tests.aswwu.behaviors.mask.mask_data import BASE_PROFILE, IMPERSONAL_FIELDS, PERSONAL_FIELDS, SELF_FIELDS
 import tests.utils as utils
+from tests.aswwu.behaviors.mask.mask_utils import build_profile_dict, assert_update_profile
 from tests.aswwu.data.paths import USERS_PATH
 from tests.aswwu.behaviors.auth.auth_subtests import assert_verify_login
 from tests.aswwu.behaviors.mask import mask_requests
 import json
 from settings import testing
-
-
-BASE_PROFILE = {
-        u'attached_to': u'Some Name',
-        u'birthday': u'01-01',
-        u'career_goals': u'Web Development',
-        u'class_of': u'2020',
-        u'class_standing': u'Senior',
-        u'department': u'None',
-        u'email': u'armanda.Woolston@wallawalla.edu',
-        u'favorite_books': u'Clean Architecture',
-        u'favorite_food': u'Byte size snacks',
-        u'favorite_movies': u'Some Movie',
-        u'favorite_music': u'Some Music',
-        u'gender': u'Other',
-        u'graduate': u'No',
-        u'high_school': u'SomePlace High',
-        u'hobbies': u'Test Driven Development',
-        u'majors': u'Computer Science',
-        u'minors': u'Web Design and Development',
-        u'office': u'None',
-        u'office_hours': u'None',
-        u'personality': u'ABCD',
-        u'pet_peeves': u'Broken Code',
-        u'phone': u'111-111-1111',
-        u'photo': u'images/default_mask/default.jpg',
-        u'preprofessional': u'No',
-        u'privacy': u'1',
-        u'quote': u'A looooooooooooooooooooooooooooooooooooooooooooooooo'
-                  u'ooooooooooooooooooooooooooooonnnnnnnnnnnnnnnnnnnnnnn'
-                  u'nnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnnngg'
-                  u'ggggggggggggggggggggggg string of text with some wei'
-                  u'rd chars !@#$%&amp;*()_+-=`~?&gt;&lt;|}{\\][\';": an'
-                  u'd a second " plus a newline\n\nand a\t tab',
-        u'quote_author': u'Stephen Ermshar',
-        u'relationship_status': u'In engineering...',
-        u'website': u'aswwu.com',
-    }
-
-# fields that should be visible regardless of the viewer's status
-BASE_FIELDS = {
-    "email",  # email is visible to not logged in users for individuals with privacy==0, but not for privacy==1, yikes..
-    "full_name",
-    "photo",
-    "username",
-    "views",
-}
-
-# fields that should only be visible to logged out users if the profile's privacy is "1"
-IMPERSONAL_FIELDS = {
-    'career_goals',
-    'department',
-    'favorite_books',
-    'favorite_movies',
-    'favorite_music',
-    'gender',
-    'graduate',
-    'hobbies',
-    'majors',
-    'minors',
-    'office',
-    'office_hours',
-    'personality',
-    'pet_peeves',
-    'preprofessional',
-    'privacy',
-    'quote',
-    'quote_author',
-    'relationship_status',
-    'website',
-}
-
-# fields that should only be visible to logged in users
-PERSONAL_FIELDS = {
-    'attached_to',
-    'birthday',
-    'class_of',
-    'class_standing',
-    'high_school',
-    'phone',
-}
-
-# fields that should only be visible to a profile's owner
-SELF_FIELDS = {
-    "wwuid",
-    "favorite_food",  # change this, these regression tests are being written for current behavior, not desired
-}
-
-
-def build_profile_dict(user, update_dict=dict(), remove_keys=set(), base_dict=BASE_PROFILE):
-    """
-    given a user, start with the BASE_PROFILE, or base_dict if provided, and construct a user object,
-    adding or updating fields in update_dict, and removing fields listed in remove keys.
-    :param user: a user for which to build the profile
-    :param update_dict: (optional) a python dictionary of items to be added or updated
-    :param remove_keys: (optional) a python set of string keys to be removed
-    :param base_dict: (optional) an alternative base profile dictionary
-    :return:
-    """
-    new_profile = dict()
-    new_profile.update(base_dict)
-    new_profile.update(user)
-    new_profile.update(update_dict)
-    remove_keys = remove_keys.intersection(set(new_profile.keys()))
-    for remove_key in remove_keys:
-        del new_profile[remove_key]
-    return new_profile
-
-
-def assert_update_profile(user, session, profile_dict):
-    update_response = mask_requests.post_update_profile(user["username"], profile_dict, session)
-    assert update_response.status_code == 200
-    assert json.loads(update_response.text) == "success"
 
 
 def test_update_profile(testing_server):
